fix: Footer 회사 정보
