+++ src/components/Footer/index.jsx
@@ -1,78 +1,12 @@
-import { useNavigate } from "react-router-dom";
+import Company from "./Company";
+import Policy from "./Policy";
 
 function Footer() {
-  const navigation = useNavigate();
-
   return (
-    <div className="grid mt-5 mb-24 md:mb-5 px-3 font-medium text-xs text-slate-600">
-      <div className="hidden sm:flex gap-3 justify-center items-center flex-wrap">
-        <button
-          className="hover:underline underline-offset-2 focus:outline-none"
-          onClick={() => navigation("/intro")}
-        >
-          회사 소개
-        </button>
-        <span>|</span>
-        <button
-          className="hover:underline underline-offset-2 focus:outline-none"
-          onClick={() => navigation("/policy/user/service")}
-        >
-          이용약관
-        </button>
-        <span>|</span>
-        <button
-          className="hover:underline underline-offset-2 text-primary focus:outline-none"
-          onClick={() => navigation("/policy/user/privacy")}
-        >
-          개인정보처리방침
-        </button>
-        <span>|</span>
-        <button
-          className="hover:underline underline-offset-2 focus:outline-none"
-          onClick={() => navigation("/policy/user/location")}
-        >
-          위치기반 서비스 이용약관
-        </button>
-        <span>|</span>
-        <span>통신판매업신고번호 2024-경기의왕-0056호</span>
-      </div>
-      <div className="sm:hidden">
-        <div className="flex gap-3 justify-center items-center flex-wrap">
-          <button
-            className="hover:underline underline-offset-2 focus:outline-none"
-            onClick={() => navigation("/intro")}
-          >
-            회사 소개
-          </button>
-          <span>|</span>
-          <button
-            className="hover:underline underline-offset-2 focus:outline-none"
-            onClick={() => navigation("/policy/user/service")}
-          >
-            이용약관
-          </button>
-          <span>|</span>
-          <button
-            className="hover:underline underline-offset-2 text-primary focus:outline-none"
-            onClick={() => navigation("/policy/user/privacy")}
-          >
-            개인정보처리방침
-          </button>
-        </div>
-        <div className="mt-2 flex gap-3 justify-center items-center flex-wrap">
-          <button
-            className="hover:underline underline-offset-2 focus:outline-none"
-            onClick={() => navigation("/policy/user/location")}
-          >
-            위치기반 서비스 이용약관
-          </button>
-          <span>|</span>
-          <span>통신판매업신고번호 2024-경기의왕-0056호</span>
-        </div>
-      </div>
-      <div className="flex justify-center items-center pt-3">
-        Copyright ⓒ 2024 mallangtrip. All rights reserved.
-      </div>
+    <div className="cursor-default max-w-screen-xl mx-auto">
+      <hr className="bg-darkgray/20 mt-14 mb-7 h-px border-0 mx-5" />
+      <Company />
+      <Policy />
     </div>
   );
 }
