+++ src/components/LoginLayout/index.jsx
@@ -1,12 +1,10 @@
 import { Outlet } from "react-router-dom";
 import Footer from "../Footer";
-import BottomInfo from "../../pages/IntroPage/BottomInfo";
 
 function LoginLayout() {
   return (
     <>
       <Outlet />
-      <BottomInfo />
       <Footer />
     </>
   );
